+++ modules/modules_manager/src/cpp/ModulesManager.cpp
@@ -222,18 +222,14 @@ ModulesManager::deleteAllModules()
 bool
 ModulesManager::deleteModule(const std::wstring& modulename)
 {
-    bool found = false;
-    size_t pos = 0;
-    for (size_t i = 0; i < modulesMap.size(); i++) {
-        mapElement elem = modulesMap[i];
-        if (std::get<0>(elem) == modulename) {
-            pos = i;
-            found = true;
+    std::vector<mapElement>::const_iterator it;
+    for (it = modulesMap.begin(); it != modulesMap.end(); ++it) {
+        if (std::get<0>(*it) == modulename) {
             break;
         }
     }
-    if (found) {
-        modulesMap.erase(modulesMap.begin() + pos);
+    if (it != modulesMap.end()) {
+        modulesMap.erase(it);
         return true;
     }
     return false;
@@ -242,19 +238,14 @@ ModulesManager::deleteModule(const std::wstring& modulename)
 bool
 ModulesManager::findModule(const std::wstring& modulename, std::wstring& path)
 {
-    bool found = false;
-    size_t pos = 0;
-    for (size_t i = 0; i < modulesMap.size(); i++) {
-        mapElement elem = modulesMap[i];
-        if (std::get<0>(elem) == modulename) {
-            pos = i;
-            found = true;
+    std::vector<mapElement>::const_iterator it;
+    for (it = modulesMap.begin(); it != modulesMap.end(); ++it) {
+        if (std::get<0>(*it) == modulename) {
             break;
         }
     }
-    if (found) {
-        mapElement elem = modulesMap[pos];
-        path = std::get<1>(elem);
+    if (it != modulesMap.end()) {
+        path = std::get<1>(*it);
         return true;
     }
     return false;