+++ modules/os_functions/src/cpp/GetUsername.cpp
@@ -1,0 +1,59 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// Alternatively, you can redistribute it and/or
+// modify it under the terms of the GNU General Public License as
+// published by the Free Software Foundation; either version 2 of
+// the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this program. If not, see <http://www.gnu.org/licenses/>.
+// LICENCE_BLOCK_END
+//=============================================================================
+#ifdef _MSC_VER
+#include <Windows.h>
+#endif
+#include "GetUsername.hpp"
+#include "GetVariableEnvironment.hpp"
+//=============================================================================
+namespace Nelson {
+//=============================================================================
+static std::wstring username;
+//=============================================================================
+std::wstring
+GetUsername()
+{
+    if (username.empty()) {
+#ifdef _MSC_VER
+#define INFO_BUFFER_SIZE 32767
+        TCHAR infoBuf[INFO_BUFFER_SIZE];
+        DWORD bufCharCount = INFO_BUFFER_SIZE;
+        if (!GetUserName(infoBuf, &bufCharCount)) {
+            username = GetVariableEnvironment(L"USERNAME");
+        } else {
+            username = infoBuf;
+        }
+#else
+        username = GetVariableEnvironment(L"USER");
+        if (username.empty()) {
+            username = GetVariableEnvironment(L"LOGNAME");
+        }
+#endif
+    }
+  return username;
+}
+//=============================================================================
+}