+++ modules/types/src/include/nlsConfig.h
@@ -37,15 +37,15 @@
 #ifdef _MSC_VER
 #ifdef _DEBUG
 #ifdef _WIN64
-#define BOOST_TARGET "vc142-mt-gd-x64-1_73"
+#define BOOST_TARGET "vc142-mt-gd-x64-1_74"
 #else
-#define BOOST_TARGET "vc142-mt-gd-x32-1_73"
+#define BOOST_TARGET "vc142-mt-gd-x32-1_74"
 #endif
 #else
 #ifdef _WIN64
-#define BOOST_TARGET "vc142-mt-x64-1_73"
+#define BOOST_TARGET "vc142-mt-x64-1_74"
 #else
-#define BOOST_TARGET "vc142-mt-x32-1_73"
+#define BOOST_TARGET "vc142-mt-x32-1_74"
 #endif
 #endif
 #endif